requirements.txt
babel
chart-studio
numpy
plotly
pycountry
python-dotenv
pytz
wolframalpha